requirements.txt
icalendar>=6.0.0
requests>=2.31.0
pyyaml>=6.0.1
jinja2>=3.1.3
anthropic>=0.40.0
openai>=1.50.0
google-genai>=1.0.0
tqdm>=4.66.0
